:seedling: Use `cross-env` to let npm scripts work on Windows
Bash style of setting an env var for a command doesn't
work on a Windows shell.  So instead of forcing Windows
users to install bash and reconfigure npm to use it,
run scripts with `cross-env`.  This should allow the
same scripts to work everywhere.

For reference, to setup npm to use bash that is included
with the git for windows installation, this command
should work:
```
npm config set script-shell "C:\\Program Files\\Git\\bin\\bash.exe
```

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>